Split DbDescribeCest into 3 adapter cests.
